requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
mysqlclient==2.1.1
pytz==2022.1
sqlparse==0.4.2
style==1.1.0
tzdata==2022.1
update==0.0.1
